+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
-## [1.0.0] - 2022-09-22
+## [1.0.1] - 2022-09-26
+
+### Added
+
+- Vanilla .toc file
+
+### Changed
+
+- Windows are now added to UISpecialFrames
+- Option buttons now dynamically resize
+- Certain UI interactions now play sounds
+
+## [1.0.0] - 2022-09-25
 
 ### Changed